+++ imports/charts/client/index.js
@@ -1,8 +1,10 @@
 import './message-day-chart';
 
-Router.route('/charts', function() {
-  this.layout('ApplicationLayout');
-  this.render('MessageDayChart');
-}, {
-  name: 'charts'
+ChartsController = ApplicationController.extend({
+  template: 'MessageDayChart'
+});
+
+Router.route('/charts', {
+  name: 'charts',
+  controller: 'ChartsController'
 });